relative frame

+++ src/game_data.rs
@@ -55,9 +55,9 @@ pub struct GameData {
     pub das: usize,
     pub arr: usize,
     pub soft_drop_gravity: f32,
-    pub freeze_delay: usize,
+    pub freeze_delay: f32,
 
-    pub freeze_left: usize,
+    pub freeze_left: f32,
     pub accumulated_down: f32,
     pub accumulated_move: f32,
 
@@ -78,10 +78,10 @@ impl GameData {
             gravity: 0.0156,
             das: 9,
             arr: 2,
-            freeze_delay: 30,
+            freeze_delay: 30.,
             soft_drop_gravity: 20.,
 
-            freeze_left: 30,
+            freeze_left: 30.,
             accumulated_down: 0.,
             accumulated_move: 0.,
